web/backend/Python-services/requirements.txt
pymongo==4.4.1
python-dotenv==1.0.0
pydantic==2.6.1
click==8.1.7
fastapi==0.100.0
uvicorn==0.23.1
bcrypt==4.0.1
python-dateutil==2.8.2
python-multipart==0.0.6
numpy==1.24.4
pandas==2.0.3
openpyxl==3.1.2
requests==2.31.0
Pillow==10.1.0 